# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExtraWork API",
    "version": "1.0.0",
    "description": "WhatsApp-first contract-change and approval ledger.\n\nThree surfaces: `/v1` (authenticated business), `/public/v1` (no-account customer approval), `/webhooks/v1` (provider callbacks). Money is always integer minor units. Quantities are decimal strings. Instants are ISO-8601 UTC.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment"
    }
  ],
...
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "MALFORMED_JSON",
                  "PAYLOAD_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "MISSING_IDEMPOTENCY_KEY",
                  "MISSING_IF_MATCH",
                  "UNAUTHENTICATED",
                  "SESSION_EXPIRED",
                  "REAUTHENTICATION_REQUIRED",
                  "FORBIDDEN",
                  "CSRF_FAILED",
                  "ORGANIZATION_REQUIRED",
                  "NOT_A_MEMBER",
                  "ORGANIZATION_SUSPENDED",
                  "NOT_FOUND",
                  "CUSTOMER_NOT_FOUND",
                  "PROJECT_NOT_FOUND",
                  "CHANGE_ORDER_NOT_FOUND",
                  "CONTACT_NOT_FOUND",
                  "ATTACHMENT_NOT_FOUND",
                  "EMPLOYEE_NOT_FOUND",
                  "INVALID_STATE_TRANSITION",
                  "ALREADY_DECIDED",
                  "VERSION_SUPERSEDED",
                  "REQUEST_EXPIRED",
                  "TOKEN_REVOKED",
                  "TOKEN_INVALID",
                  "LOCK_CONFLICT",
                  "IDEMPOTENCY_KEY_REUSED",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "ETAG_MISMATCH",
                  "CURRENCY_MISMATCH",
                  "NEGATIVE_REVISED_TOTAL",
                  "EMPTY_CHANGE",
                  "BASELINE_LOCKED",
                  "PROJECT_CLOSED",
                  "PROJECT_INTEGRITY_REVIEW",
                  "ATTACHMENT_NOT_READY",
                  "ATTACHMENT_IMMUTABLE",
                  "DUPLICATE_NUMBER",
                  "EMPLOYEE_PHONE_TAKEN",
                  "ASSURANCE_REQUIRED",
                  "ASSURANCE_UNAVAILABLE",
                  "OTP_INVALID",
                  "OTP_EXPIRED",
                  "OTP_ATTEMPTS_EXCEEDED",
                  "ENTITLEMENT_EXCEEDED",
                  "SUBSCRIPTION_READ_ONLY",
                  "FEATURE_NOT_ENTITLED",
                  "FILE_TYPE_NOT_ALLOWED",
                  "FILE_TOO_LARGE",
                  "FILE_SCAN_FAILED",
                  "FILE_CONTENT_MISMATCH",
                  "RATE_LIMITED",
                  "WEBHOOK_SIGNATURE_INVALID",
                  "PROVIDER_UNAVAILABLE",
                  "NOT_IMPLEMENTED",
                  "INTERNAL_ERROR",
                  "SERVICE_UNAVAILABLE"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "ExtraWorkReport": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "changeOrderId": {
                  "type": "string",
                  "format": "uuid"
                },
                "number": {
                  "type": "string"
                },
                "versionNumber": {
                  "type": "integer"
                },
                "projectId": {
                  "type": "string",
                  "format": "uuid"
                },
                "projectNumber": {
                  "type": "string"
                },
                "projectTitle": {
                  "type": "string"
                },
                "customerName": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "currency": {
                  "type": "string"
                },
                "subtotalDeltaMinor": {
                  "type": "integer"
                },
                "taxDeltaMinor": {
                  "type": "integer"
                },
                "totalDeltaMinor": {
                  "type": "integer"
                },
                "scheduleDeltaDays": {
                  "type": "integer"
                },
                "sentAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "ISO 8601 instant in UTC",
                  "nullable": true
                },
                "decidedAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "ISO 8601 instant in UTC",
                  "nullable": true
                },
                "decisionType": {
                  "type": "string",
                  "nullable": true
                },
                "assuranceAchieved": {
                  "type": "string",
                  "nullable": true
                },
                "createdBy": {
                  "type": "string"
                }
              },
              "required": [
                "changeOrderId",
                "number",
                "versionNumber",
                "projectId",
                "projectNumber",
                "projectTitle",
                "customerName",
                "title",
                "status",
                "currency",
                "subtotalDeltaMinor",
                "taxDeltaMinor",
                "totalDeltaMinor",
                "scheduleDeltaDays",
                "sentAt",
                "decidedAt",
                "decisionType",
                "assuranceAchieved",
                "createdBy"
              ],
              "additionalProperties": false
            }
          },
          "totals": {
            "type": "object",
            "properties": {
              "currency": {
                "type": "string"
              },
              "count": {
                "type": "integer"
              },
              "subtotalDeltaMinor": {
                "type": "integer"
              },
              "taxDeltaMinor": {
                "type": "integer"
              },
              "totalDeltaMinor": {
                "type": "integer"
              }
            },
            "required": [
              "currency",
              "count",
              "subtotalDeltaMinor",
              "taxDeltaMinor",
              "totalDeltaMinor"
            ],
            "additionalProperties": false
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 instant in UTC"
          }
        },
        "required": [
          "rows",
          "totals",
          "generatedAt"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ew_session",
        "description": "HTTP-only session cookie. Mutations additionally require the X-CSRF-Token header matching the ew_csrf cookie (report §6.5, §12.1)."
      },
      "publicSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ew_public",
        "description": "Short-lived public approval session, created when a valid link is first opened."
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/v1/projects/{id}/evidence-pack": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Request a project evidence pack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Project id"
          }
        ],
        "responses": {
          "202": {
            "description": "Queued"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: state, lock, or idempotency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{id}/accounting-export": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Approved changes in the accounting hand-off model",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Project id"
          }
        ],
        "responses": {
          "200": {
            "description": "Export model"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: state, lock, or idempotency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reports/extra-work": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Documented extra-work summary",
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraWorkReport"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: state, lock, or idempotency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/extra-work.csv": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Extra-work summary as CSV",
        "responses": {
          "200": {
            "description": "CSV"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: state, lock, or idempotency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exports/{id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export job status and signed download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Export id"
          }
        ],
        "responses": {
          "200": {
            "description": "Export"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: state, lock, or idempotency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}